convert js into jsx
