+++ babelrc
@@ -1,0 +1,7 @@
+{
+    "plugins": [
+      ["@babel/plugin-transform-react-jsx", {
+        "runtime": "automatic"
+      }]
+    ]
+  }
