+++ src/Components/dashboard.jsx
@@ -14,7 +14,6 @@ import {
   Col,
   Dropdown,
 } from "react-bootstrap";
-
 import { useDropzone } from "react-dropzone";
 import { useNavigate } from "react-router-dom";
 import * as XLSX from "xlsx";
@@ -30,14 +29,15 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "../styles/dashboard.css";
-
+import PopUpContainer from "./popup";
+import ResetPassword from "./resetPassword";
 import CustomSnackbar from "./Snackbar";
 import { PortURL } from "./Config";
-import NavbarSection from "./NavbarSection";
+
+
 
 function Dashboard() {
   const [username, setUsername] = useState("");
-  
   const [data, setData] = useState([]);
   const [searchQuery, setSearchQuery] = useState("");
   const [editedRowId, setEditedRowId] = useState(null); // Track edited row ID
@@ -50,9 +50,11 @@ function Dashboard() {
   const [snackbarOpen, setSnackbarOpen] = useState(false);
   const [retriveData, setRetriveData] = useState([]);
   const [snackbarMessage, setSnackbarMessage] = useState("");
+  const [isMobile, setIsMobile] = useState(false);
 
   const navigate = useNavigate();
 
+  
   useEffect(() => {
     const isLoggedIn = localStorage.getItem("isLoggedIn");
     if (!isLoggedIn) {
@@ -67,6 +69,18 @@ function Dashboard() {
     }
   }, [navigate]);
 
+  useEffect(() => {
+    const handleResize = () => {
+      setIsMobile(window.innerWidth >= 1000);
+    };
+
+    handleResize();
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
 
   useEffect(() => {
     if (uploadSuccess) {
@@ -146,28 +160,27 @@ function Dashboard() {
     setEditedRowData(filteredData[rowId]);
   };
 
-
   const handleCancel = () => {
     console.log("Canceling edit for row:", editedRowId);
     setEditedRowId(null);
   };
 
-const handleInputChange = (e, key) => {
-  const { value } = e.target;
-  console.log("Key:", key);
-  console.log("Value:", value);
-  // Update the edited row data with the new value
-  setEditedRowData((prevData) => ({
-    ...prevData,
-    [String(key)]: String(value || "") // Convert both key and value to strings and provide a fallback value of an empty string if value is null or undefined
-  }));
-};
+  const handleInputChange = (e, key) => {
+    const { value } = e.target;
+    console.log("Key:", key);
+    console.log("Value:", value);
+    // Update the edited row data with the new value
+    setEditedRowData((prevData) => ({
+      ...prevData,
+      [String(key)]: String(value || ""), // Convert both key and value to strings and provide a fallback value of an empty string if value is null or undefined
+    }));
+  };
 
-  
-  
-  
-
-
+  const handleLogout = () => {
+    localStorage.removeItem("isLoggedIn");
+    localStorage.removeItem("username");
+    navigate("/login");
+  };
 
   const handleInvite = () => {
     navigate("/send-invite");
@@ -251,18 +264,21 @@ const handleInputChange = (e, key) => {
     }
   };
 
-
   const handleSave = async () => {
     try {
       console.log("Edit saved data", editedRowData);
-      
-       
+
       // Format the MonthYear date to "YYYY-MM-DD"
       const monthYearDate = new Date(editedRowData.MonthYear);
-      const formattedMonthYear = `${monthYearDate.getFullYear()}-${(monthYearDate.getMonth() + 1)
+      const formattedMonthYear = `${monthYearDate.getFullYear()}-${(
+        monthYearDate.getMonth() + 1
+      )
         .toString()
-        .padStart(2, "0")}-${monthYearDate.getDate().toString().padStart(2, "0")}`;
-  
+        .padStart(2, "0")}-${monthYearDate
+        .getDate()
+        .toString()
+        .padStart(2, "0")}`;
+
       // Create the payload with the updated MonthYear format
       const payload = {
         editedRow: {
@@ -270,7 +286,7 @@ const handleInputChange = (e, key) => {
           MonthYear: formattedMonthYear,
         },
       };
-  
+
       // Send the updated data to the server
       const response = await fetch(`${PortURL}/update`, {
         method: "POST", // Change the method to POST since you're sending data
@@ -279,7 +295,7 @@ const handleInputChange = (e, key) => {
         },
         body: JSON.stringify(payload), // Send only the edited row data
       });
-  
+
       if (response.ok) {
         console.log("Row updated successfully:", editedRowData);
         // Optionally, you can refresh the data from the server
@@ -298,14 +314,13 @@ const handleInputChange = (e, key) => {
       setSnackbarMessage("Error updating row");
     }
   };
-  
 
   const handleDelete = async (rowId) => {
     try {
       console.log("Row ID:", rowId);
-    
+
       console.log("Filtered Data:", filteredData);
-      
+
       const identifierToDelete = String(filteredData[rowId]?.ID); // Convert identifierToDelete to string
       console.log("Identifier to Delete:", identifierToDelete);
 
@@ -333,7 +348,6 @@ const handleInputChange = (e, key) => {
       setSnackbarMessage("Error deleting row");
     }
   };
-  
 
   const handleCloseSnackbar = () => {
     setSnackbarOpen(false);
@@ -342,22 +356,66 @@ const handleInputChange = (e, key) => {
 
   return (
     <div className="dashboard-container">
+      <Navbar bg="light" expand="lg" className="w-100">
+        <div className="brand-wrapper">
+          <NavbarBrand href="#home">
+            <img src="/images/bcp2.png" alt="Logo" className="customLogo" />
+          </NavbarBrand>
+        </div>
+        <NavbarToggle aria-controls="basic-navbar-nav" />
+        <NavbarCollapse id="basic-navbar-nav">
+          <Nav className="ml-auto align-items-center">
+            {/* Render username inside dropdown menu if isMobile is true */}
+            {isMobile ? (
+              <Dropdown className="d-flex">
+                <Dropdown.Toggle
+                  id="dropdown-basic"
+                  as="div"
+                  className="customDropdown"
+                >
+                  <FontAwesomeIcon className="username " icon={faUser}  /> USERNAME
+                </Dropdown.Toggle>
+                <Dropdown.Menu>
+                  <PopUpContainer>
+                    <ResetPassword  />
+                  </PopUpContainer>
+                  <Dropdown.Item onClick={handleLogout}>Logout</Dropdown.Item>
+                </Dropdown.Menu>
+              </Dropdown>
+            ) : (
+              // Render the PopUpContainer directly if not mobile
+              
+             
+              <React.Fragment >
+                <div className="smallscreen">
 
-      <NavbarSection/>
+              <div className="ml-auto align-items-center user ">
+                <FontAwesomeIcon icon={faUser} /> {username}
+              </div>
+              <PopUpContainer className="popUp" >
+                <ResetPassword className="popUp" />
+              </PopUpContainer> 
+              <Dropdown.Item onClick={handleLogout} className="logout">Logout</Dropdown.Item>
+              </div>
+
+            </React.Fragment>
+            )}
+          </Nav>
+        </NavbarCollapse>
+      </Navbar>
+
       <Container fluid>
         <div className="container-fluid full-height mt-5">
           <div className="row">
             <div className="col">
-              <div className="border shadow p-3 d-flex justify-content-between align-items-center">
-               
-               
+              <div className="border shadow p-3 d-flex justify-content-between align-items-center ">
                 <Form className="d-flex ">
                   <div className="search-wrapper mr-2">
                     {/* <div className="search-icon">
                       <FontAwesomeIcon icon={faSearch} />
                     </div> */}
                     <FormControl
-                    className="search-input"
+                      className="search-input"
                       type="text"
                       placeholder="Search"
                       style={{ flex: "1" }}
@@ -365,26 +423,30 @@ const handleInputChange = (e, key) => {
                       onChange={handleSearchChange}
                     />
                   </div>
-                  
+
                   <div {...getRootProps()} className="custom-file-upload ">
-                    <input {...getInputProps()}  accept=".xlsx, .xls" />
+                    <input {...getInputProps()} accept=".xlsx, .xls" />
                     {isDragActive ? (
                       <p>Drop the files here ...</p>
                     ) : (
                       <Button className="btn btn-secondary btn-sm upload">
-                        <FontAwesomeIcon icon={faUpload}  />
+                        <FontAwesomeIcon icon={faUpload} />
                         Upload File
                       </Button>
                     )}
                   </div>
                 </Form>
                 <div className="ml-4 ">
-                  <Button className="mr-2 btn btn-secondary submit" onClick={handleSubmit}>
+                  <Button
+                    className="mr-2 btn btn-secondary submit"
+                    onClick={handleSubmit}
+                  >
                     Submit
                   </Button>
                   <Button className="clear" variant="danger">
                     {" "}
-                    Clear <FontAwesomeIcon icon={faTrash} />
+                    Clear{" "}
+                    <FontAwesomeIcon className="clearicon" icon={faTrash} />
                   </Button>
                 </div>
               </div>
@@ -398,9 +460,9 @@ const handleInputChange = (e, key) => {
         message={snackbarMessage}
         onClose={handleCloseSnackbar}
       />
-      <Container fluid>
-        <Row>
-          <Col>
+      <Container fluid >
+        <Row className="row Render-Row">
+          <Col className="col Render-Col">
             <div className="table-responsive render">
               <Table striped bordered hover>
                 <thead>
@@ -419,7 +481,7 @@ const handleInputChange = (e, key) => {
                   </tr>
                 </thead>
                 <tbody>
-                                  {filteredData.map((row, index) => (
+                  {filteredData.map((row, index) => (
                     <tr key={index}>
                       <td className="selection-cell">
                         <input
@@ -429,54 +491,53 @@ const handleInputChange = (e, key) => {
                         />
                       </td>
                       {Object.keys(row).map((key) => (
-  <td key={key}>
-    {editedRowId === index ? (
-      <input
-        type="text"
-        value={editedRowData[key] || ""}
-        onChange={(e) => handleInputChange(e, key)}
-      />
-    ) : (
-      formatDateCell(row[key], key)
-    )}
-  </td>
-))}
-<td className="action-cell">
-  {editedRowId === index ? (
-    <div className="action-buttons">
-      <button
-        className="btn btn-primary btn-sm"
-        onClick={() => handleSave()}
-      >
-        <FontAwesomeIcon icon={faSave} />
-      </button>
-      <button
-        className="btn btn-secondary btn-sm"
-        onClick={() => handleCancel()}
-      >
-        <FontAwesomeIcon icon={faTimes} />
-      </button>
-    </div>
-  ) : (
-    <div className="action-buttons">
-      <button
-        className="btn btn-primary btn-sm"
-        onClick={() => handleEdit(index)}
-      >
-        <FontAwesomeIcon icon={faEdit} />
-      </button>
-      <button
-        className="btn btn-danger btn-sm"
-        onClick={() => handleDelete(index)}
-      >
-        <FontAwesomeIcon icon={faTrash} />
-      </button>
-    </div>
-  )}
-</td>
+                        <td key={key}>
+                          {editedRowId === index ? (
+                            <input
+                              type="text"
+                              value={editedRowData[key] || ""}
+                              onChange={(e) => handleInputChange(e, key)}
+                            />
+                          ) : (
+                            formatDateCell(row[key], key)
+                          )}
+                        </td>
+                      ))}
+                      <td className="action-cell">
+                        {editedRowId === index ? (
+                          <div className="action-buttons">
+                            <button
+                              className="btn  btn-sm Save"
+                              onClick={() => handleSave()}
+                            >
+                              <FontAwesomeIcon icon={faSave} />
+                            </button>
+                            <button
+                              className="btn btn-sm Cancel"
+                              onClick={() => handleCancel()}
+                            >
+                              <FontAwesomeIcon icon={faTimes} />
+                            </button>
+                          </div>
+                        ) : (
+                          <div className="action-buttons">
+                            <button
+                              className="btn  btn-sm Edit"
+                              onClick={() => handleEdit(index)}
+                            >
+                              <FontAwesomeIcon icon={faEdit} />
+                            </button>
+                            <button
+                              className="btn btn-sm Delete"
+                              onClick={() => handleDelete(index)}
+                            >
+                              <FontAwesomeIcon icon={faTrash} />
+                            </button>
+                          </div>
+                        )}
+                      </td>
                     </tr>
                   ))}
-                    
                 </tbody>
               </Table>
             </div>
