+++ src/styles/dashboard.css
@@ -5,8 +5,8 @@
 .navbar {
   position: sticky;
   top: 0;
-  z-index: 1000; /* Ensure it stays above other content */
-  background-color: #ffffff; /* Add a background color */
+  z-index: 1000; 
+  background-color: #ffffff; 
   border-bottom: 1px solid #dee2e6;
 }
 
@@ -18,11 +18,7 @@
   background-color: none;
 }
 
-/* .dropdown-toggle {
-  background-color: transparent !important;
-  border: none;
-  color: black;
-} */
+
 
 .dropdown-toggle,
 .dropdown-menu {
@@ -37,12 +33,10 @@
 .form-inline {
   display: flex;
   gap: 10px;
-  /* Add a gap between the search bar and the upload button */
 }
 
 .brand-wrapper {
   padding: 0 15px;
-  /* Adjust the padding as needed */
 }
 
 .search-wrapper {
@@ -52,14 +46,11 @@
 
 .search-icon {
   margin-right: 10px;
-  /* Adjust the margin as needed */
 }
 
-/* dashboard.css */
 
 .table-container {
   box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-  /* Gmail-like box shadow */
 }
 
 .table-container table {
@@ -100,13 +91,13 @@
 
 button {
   width: 150px;
-  height: 40px; /* Set height for all buttons */
+  height: 40px; 
  
 }
 
 .dropdown-menu .dropdown-item:hover {
   color: inherit; 
-  background-color: transparent; /* Remove background color change */
+  background-color: transparent; 
 }
 
 .custom-file-upload{
@@ -118,18 +109,17 @@ button {
 }
 
 
-/* Add responsive styles here */
 .excel-grid-container {
   width: 100%;
   overflow-x: auto;
-  overflow-y: auto; /* Add vertical scrolling */
+  overflow-y: auto; 
   /* padding: 10px; */
 
 }
 
 .table {
-  min-width: 600px; /* Set a minimum width for the table to prevent collapsing */
-  width: max-content; /* Ensure the table does not exceed its container */
+  min-width: 600px; 
+  width: max-content; 
   border-collapse: collapse;
 }
 .table th,
@@ -309,7 +299,185 @@ button {
 }
 
 
+      .nav-link {
+        padding-right: 0.5rem !important;
+        padding-left: 0.5rem !important;
+      }
 
+      .dropdown-toggle {
+        margin-right: 10px;
+      }
+
+      .dropdown-menu {
+        margin-top: 5px;
+      }
+
+     
+
+      .brand-wrapper>a {
+        color: rgb(7, 106, 7);
+      }
+
+      .dropdown-basic {
+        background-color: none;
+      }
+
+      .dropdown-toggle,
+      .dropdown-menu {
+        min-width: 150px;
+      }
+
+      .navbar-nav {
+        width: 100%;
+        justify-content: flex-end;
+      }
+
+      .form-inline {
+        display: flex;
+        gap: 10px;
+
+      }
+
+      .brand-wrapper {
+        padding: 0 15px;
+
+      }
+
+      .search-wrapper {
+        display: flex;
+        align-items: center;
+      }
+
+      .search-icon {
+        margin-right: 10px;
+      }
+
+      .table-container {
+        box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1), 0px 1px 3px rgba(0, 0, 0, 0.08);
+        border-radius: 8px;
+
+      }
+
+      .table-container table {
+        border: 1px solid #dee2e6;
+        border-radius: 5px;
+      }
+
+      .table-container th,
+      .table-container td {
+        border: 1px solid #dee2e6;
+        padding: 8px;
+      }
+
+      .table-container th {
+        background-color: #efeded;
+      }
+
+      .table-container tbody tr:hover {
+        background-color: #f0f0f0;
+      }
+
+      .customLogo {
+        width: 50px;
+        height: auto;
+        max-height: 50px;
+      }
+
+      button {
+        width: 150px;
+        height: 40px;
+
+      }
+
+      .dropdown-menu .dropdown-item:hover {
+        color: inherit;
+        background-color: transparent;
+
+      }
+
+      .custom-file-upload {
+        margin-left: 10px;
+      }
+
+      .submit {
+        margin-right: 10px;
+      }
+
+      .excel-grid-container {
+        width: 100%;
+        overflow-x: auto;
+        overflow-y: auto;
+
+      }
+
+      .table {
+        min-width: 600px;
+
+        width: max-content;
+
+        border-collapse: collapse;
+      }
+
+      .table th,
+      .table td {
+        padding: 10px;
+        text-align: left;
+        white-space: nowrap;
+      }
+
+      .table th {
+        background-color: #f2f2f2;
+        border-bottom: 2px solid #ddd;
+        text-transform: uppercase
+      }
+
+      .table-striped tbody tr:nth-of-type(odd) {
+        background-color: #f9f9f9;
+      }
+
+      .btn {
+        padding: 8px 16px;
+        font-size: 14px;
+        border-radius: 4px;
+        cursor: pointer;
+      }
+
+      /* .btn-primary {
+        background-color: #007bff;
+        border-color: #007bff;
+        color: #fff;
+      } */
+
+      .btn-secondary {
+        background-color: #28a745;
+        border-color: #28a745;
+        color: #fff;
+      }
+
+      .btn-secondary:hover {
+        background-color: #28a745;
+        border-color: #28a745;
+        color: #fff;
+      }
+
+      .btn-danger {
+        background-color: #dc3545;
+        border-color: #dc3545;
+        color: #fff;
+      }
+
+      .btn-sm {
+        padding: 6px 12px;
+        font-size: 12px;
+      }
+
+      .action-buttons {
+
+        display: flex;
+        justify-content: center;
+      }
+
+    
 
       @media (max-width: 816px) {
         .excel-grid-container {
@@ -551,7 +719,7 @@ button {
           border-radius: 5px;
           background-color: white;
           box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1), 0px 1px 3px rgba(0, 0, 0, 0.08);
-          width: 160px;
+          width: 150px;
         
         }
         
@@ -585,7 +753,10 @@ button {
 
 
 
-
+              .navbar-toggler {
+          width: 50px;
+          margin-right: 20px;
+              }
 
 
 
@@ -603,6 +774,8 @@ button {
                 color: black;
               }
 
+
+              
 
       @import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap')
